refactor according to CR
