+++ packages/salesforce-adapter/src/filters/important_values_filter.ts
@@ -14,43 +14,9 @@
 * limitations under the License.
 */
 import { CORE_ANNOTATIONS } from '@salto-io/adapter-api'
-import { ImportantValues, ImportantValue } from '@salto-io/adapter-utils'
-import { collections } from '@salto-io/lowerdash'
 import { isMetadataObjectType, MetadataObjectType } from '../transformers/transformer'
 import { LocalFilterCreator } from '../filter'
 
-const { makeArray } = collections.array
-
-
-const highligheted = (value: string): ImportantValue => ({
-  value,
-  highlighted: true,
-  indexed: false,
-})
-
-const highlightedAndIndexed = (value: string): ImportantValue => (
-  {
-    value,
-    highlighted: true,
-    indexed: true,
-  }
-)
-
-const DEFAULT_IMPORTANT_VALUES: ImportantValues = [
-  highligheted('fullName'),
-  highligheted('label'),
-  highligheted('masterLabel'),
-  highligheted('name'),
-  highligheted('content'),
-  highligheted('description'),
-  highligheted('processType'),
-  highligheted('relatedLists'),
-  highligheted('layoutSections'),
-  highligheted('lwcResources'),
-  highlightedAndIndexed('apiVersion'),
-  highlightedAndIndexed('status'),
-  highlightedAndIndexed('active'),
-]
 
 const filterCreator: LocalFilterCreator = ({ config }) => ({
   name: 'salesforceImportantValuesFilter',
@@ -58,18 +24,7 @@ const filterCreator: LocalFilterCreator = ({ config }) => ({
     if (!config.fetchProfile.isFeatureEnabled('importantValues')) {
       return
     }
-    const getImportantValues = (): ImportantValues => {
-      const additionalImportantValues = makeArray(config.fetchProfile.additionalImportantValues)
-      if (additionalImportantValues.length === 0) {
-        return DEFAULT_IMPORTANT_VALUES
-      }
-      // The additional important values should override the default ones
-      const additionalImportantValuesValues = new Set(additionalImportantValues.map(value => value.value))
-      return DEFAULT_IMPORTANT_VALUES
-        .filter(importantValue => !additionalImportantValuesValues.has(importantValue.value))
-        .concat(additionalImportantValues)
-    }
-    const importantValues = getImportantValues()
+    const { importantValues } = config.fetchProfile
     const addImportantValues = (type: MetadataObjectType): void => {
       const typeFields = new Set(Object.keys(type.fields))
       const typeImportantValues = importantValues
